IN PROGRESS - BATCH-907: Integer support in ExecutionContext
simplifying ItemStreams to take advantage of EC int support

+++ spring-batch-infrastructure/src/test/java/org/springframework/batch/item/file/FlatFileItemReaderTests.java
@@ -84,7 +84,7 @@ public class FlatFileItemReaderTests {
 		reader.read();
 		reader.read();
 
-		assertEquals(2, executionContext.getLong(ClassUtils.getShortName(FlatFileItemReader.class) + ".read.count"));
+		assertEquals(2, executionContext.getInt(ClassUtils.getShortName(FlatFileItemReader.class) + ".read.count"));
 		// close input
 		reader.close(executionContext);
 
@@ -99,7 +99,7 @@ public class FlatFileItemReaderTests {
 		assertEquals("testLine4", reader.read().toString());
 
 		reader.update(executionContext);
-		assertEquals(4, executionContext.getLong(ClassUtils.getShortName(FlatFileItemReader.class) + ".read.count"));
+		assertEquals(4, executionContext.getInt(ClassUtils.getShortName(FlatFileItemReader.class) + ".read.count"));
 	}
 
 	@Test
